requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.10.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
distro==1.9.0
easyocr==1.7.2
et_xmlfile==2.0.0
executing==2.1.0
fast-alpr==0.1.2
fast-plate-ocr==0.3.0
fastapi==0.115.12
fastjsonschema==2.20.0
filelock==3.15.4
filterpy==1.4.5
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
imutils==0.5.4
inference-sdk==0.38.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.27.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.3
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
onnxruntime-gpu==1.21.0
open-image-models==0.2.1
openai==1.58.1
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.9
pillow==10.4.0
platformdirs==4.2.2
proglog==0.1.10
progress==1.6
prometheus_client==0.21.0
prompt_toolkit==3.0.47
propcache==0.3.0
protobuf==5.29.3
psutil==6.0.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==19.0.0
pycipher==0.5.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pyproject_hooks==1.2.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
quantecon==0.7.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.3.1
setuptools==75.8.2
shapely==2.0.7
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.2
streamlit==1.42.1
supervision==0.25.1
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.2.18
tinycss2==1.3.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.48.1
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.78
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.34.2
vidstab==1.7.4
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
yarl==1.18.3
yfinance==0.2.53
